+++ jabgui/src/main/java/org/jabref/gui/entryeditor/SourceTab.java
@@ -145,7 +145,7 @@ public class SourceTab extends EntryEditorTab {
                 codeArea.getModel().replace(null, caretPos, caretPos, committed);
             }
         });
-        codeArea.getStyleClass().add("bibtex-code-area");
+        codeArea.getStyleClass().addAll("bibtex-code-area", "font-size-090");
 
         codeArea.addEventFilter(KeyEvent.KEY_PRESSED, event -> CodeAreaKeyBindings.call(codeArea, event, keyBindingRepository));
         codeArea.addEventFilter(KeyEvent.KEY_PRESSED, this::listenForSaveKeybinding);