+++ jabgui/src/main/java/org/jabref/gui/externalfiles/FileSelectionPage.java
@@ -127,7 +127,7 @@ public class FileSelectionPage extends WizardPane {
         BorderPane mainLayout = new BorderPane();
 
         progressPane = new VBox(10);
-        progressPane.getStyleClass().add("file-selection-progress-pane");
+        progressPane.getStyleClass().addAll("align-center", "padding-20");
 
         ProgressIndicator progressIndicator = new ProgressIndicator();
         progressIndicator.progressProperty().bind(viewModel.progressValueProperty());